# CUDA UpsampleNearest performance improvement