--- a/onnxruntime/core/providers/cuda/tensor/upsample_impl.cu
+++ b/onnxruntime/core/providers/cuda/tensor/upsample_impl.cu
@@ -7,9 +7,8 @@
 namespace onnxruntime {
 namespace cuda {
 
-template <typename T>
-__global__ void _UpampleNearestKernel(const size_t rank,
-                                      const TArray<int64_t> input_pitches,
+template <typename T, int RANK>
+__global__ void _UpampleNearestKernel(const TArray<int64_t> input_pitches,
                                       const TArray<fast_divmod> output_div_pitches,
                                       const TArray<fast_divmod> scales_div,
                                       const T* input_data,
@@ -20,7 +19,7 @@ __global__ void _UpampleNearestKernel(const size_t rank,
   CUDA_LONG output_index = id;
 
   int div, mod;
-  for (int dim = 0; dim < rank; ++dim) {
+  for (int dim = 0; dim < RANK; ++dim) {
     output_div_pitches[dim].divmod(output_index, div, mod);
     output_index = mod;
     if (scales_div[dim].d_ != 1 && div > 0) {
@@ -161,9 +160,25 @@ void UpampleImpl(cudaStream_t stream,
                  const size_t N) {
   int blocksPerGrid = (int)(ceil(static_cast<float>(N) / GridDim::maxThreadsPerBlock));
   if (onnxruntime::UpsampleMode::NN == upsample_mode) {
-    _UpampleNearestKernel<T><<<blocksPerGrid, GridDim::maxThreadsPerBlock, 0, stream>>>(
-        rank, input_pitches, output_div_pitches, scales_div,
-        input_data, output_data, N);
+    if (rank == 4) {
+      _UpampleNearestKernel<T,4><<<blocksPerGrid, GridDim::maxThreadsPerBlock, 0, stream>>>(
+          input_pitches, output_div_pitches, scales_div,
+          input_data, output_data, N);
+    } else if (rank == 3) {
+      _UpampleNearestKernel<T,3><<<blocksPerGrid, GridDim::maxThreadsPerBlock, 0, stream>>>(
+          input_pitches, output_div_pitches, scales_div,
+          input_data, output_data, N);
+    } else if (rank == 2) {
+      _UpampleNearestKernel<T,2><<<blocksPerGrid, GridDim::maxThreadsPerBlock, 0, stream>>>(
+          input_pitches, output_div_pitches, scales_div,
+          input_data, output_data, N);
+    } else if (rank == 1) {
+      _UpampleNearestKernel<T,1><<<blocksPerGrid, GridDim::maxThreadsPerBlock, 0, stream>>>(
+          input_pitches, output_div_pitches, scales_div,
+          input_data, output_data, N);
+    } else {
+      ORT_THROW("Unsupported rank by the Upsample CUDA kernel");
+    }
   } else if (onnxruntime::UpsampleMode::LINEAR == upsample_mode && rank == 4) {
     _UpampleBilinear4DInputKernel<T><<<blocksPerGrid, GridDim::maxThreadsPerBlock, 0, stream>>>(
         input_dim2, input_pitches, output_div_pitches, scales_div,