**Description**: This PR improves the performance of the CUDA kernel for nearest neighbor upsampling. It does so by promoting a function parameter (the rank of the tensor) into a template parameter with explicit instanciations, enabling the compiler to better unroll the loop inside the kernel. It also adds the `__restrict__` keyword to the input and output buffers, telling the compiler that the pointers cannot alias, hence enabling further optimisation.

`nvprof` before the changes (on 1.7.2):
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   69.28%  69.195ms         5  13.839ms  3.8302ms  30.768ms  void onnxruntime::cuda::_UpampleNearestKernel<float>(__int64, onnxruntime::cuda::TArray<__int64, int=8>, onnxruntime::cuda::_UpampleNearestKernel<float, onnxruntime::cuda::fast_divmod, int=8>, onnxruntime::cuda::fast_divmod, float const *, onnxruntime::cuda::_UpampleNearestKernel<float, onnxruntime::cuda::fast_divmod, int=8>*, __int64)
```
`nvprof` after the changes (still on 1.7.2):
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:    4.90%  1.6331ms         5  326.61us  93.118us  722.00us  void onnxruntime::cuda::_UpampleNearestKernel<float, int=4>(onnxruntime::cuda::TArray<__int64, int=8>, onnxruntime::cuda::_UpampleNearestKernel<float, int=4, onnxruntime::cuda::fast_divmod, int=8>, onnxruntime::cuda::fast_divmod, float const *, onnxruntime::cuda::_UpampleNearestKernel<float, int=4, onnxruntime::cuda::fast_divmod, int=8>*, __int64)
```

Speed up: x42!

**Motivation and Context**
- This enables our models, exported from TensorFlow, to run at similar or better speed than when executed from within TensorFlow.
- Closes https://github.com/microsoft/onnxruntime/issues/7578.