@cschreib-ibex 2021-05-07
FYI I noticed other kernels that suffer from this problem. I don't know if you want to port a similar fix to them as well (see `_ResizeNearestMappingKernel` for example; I haven't done an exhaustive search).