@jywu-msft 2021-06-07
/azp run Windows WebAssembly CI Pipeline, orttraining-ortmodule, orttraining-ortmodule-distributed