@cschreib-ibex 2021-09-01
FYI, I was re-reading the code and found a minor bug on line 182:
```c++
} else if (onnxruntime::UpsampleMode::LINEAR) {
```

Very sorry about this... It should have been `onnxruntime::UpsampleMode::LINEAR == upsample_mode`. Probably not a big deal in practice, as `LINEAR` has value `1` hence the `else if` branch will always be followed. But it does mean that unsupported upsample modes (like `CUBIC`) will silently behave as `LINEAR` instead of doing nothing (previous behavior). I'm not sure that either behavior is correct; but I leave that up to you.