@hariharans29 2021-09-01
> FYI, I was re-reading the code and found a minor bug on line 182:
> 
> ```c++
> } else if (onnxruntime::UpsampleMode::LINEAR) {
> ```
> 
> Very sorry about this... It should have been `onnxruntime::UpsampleMode::LINEAR == upsample_mode`. Probably not a big deal in practice, as `LINEAR` has value `1` hence the `else if` branch will always be followed. But it does mean that unsupported upsample modes (like `CUBIC`) will silently behave as `LINEAR` instead of doing nothing (previous behavior). I'm not sure that either behavior is correct; but I leave that up to you.

Thanks for catching that. 

1) Can you please fix the condition ? 

2) `Upsample` got superseded by `Resize` in opset-10 onwards. The code in the file you are modifying can only be hit by the `Upsample` kernel (and not `Resize`). Looking at the spec of [Upsample](https://github.com/onnx/onnx/blob/master/docs/Changelog.md#Upsample-9), it doesn't seem to have any other modes other than `nearest` and `linear`. But could you please go ahead and add an ORT_THROW in case we encounter some other mode in that function ? It will help others trying to understand the code in the future.